Fix circular dep in models for authorization (#723)
