+++ packages/models/src/authorization/authorizationReadModelAdapter.ts
@@ -2,8 +2,7 @@ import {
   ClientReadModel,
   KeyReadModel,
 } from "../read-models/authorizationReadModel.js";
-import { Client } from "./client.js";
-import { Key } from "./key.js";
+import { Client, Key } from "./client.js";
 
 export const toReadModelKey = (key: Key): KeyReadModel => ({
   ...key,
