+++ packages/models/src/authorization/protobufConverterToV1.ts
@@ -9,7 +9,7 @@ import {
 } from "../gen/v1/authorization/client.js";
 import { dateToBigInt } from "../utils.js";
 import { PurposeId, generateId } from "../brandedIds.js";
-import { Key, KeyUse, keyUse } from "./key.js";
+import { Key, KeyUse, keyUse } from "./client.js";
 import { Client, ClientKind, clientKind } from "./client.js";
 
 const clientComponentState = {
